[REFACTOR] category 수정

+++ src/categories/categories.controller.ts
@@ -1,9 +1,8 @@
-import { Body, Controller, Delete, Get, Param, Post, Put, Request, UseGuards } from '@nestjs/common';
+import { Body, Controller, Delete, Get, Param, Post, Request, UnauthorizedException, UseGuards } from '@nestjs/common';
 import { CategoriesService } from './categories.service';
 import { Category } from './entities/categories.entity';
 import { CreateCategoryRequestDto } from './dtos/create-categories.dto';
 import { JwtAuthGuard } from 'src/auth/jwt-auth.guard';
-import { UpdateCategoryRequestDto } from './dtos/update-categories.dto';
 
 @UseGuards(JwtAuthGuard)
 @Controller('categories')
@@ -18,27 +17,24 @@ export class CategoriesController {
   }
 
   @Post()
-  async createCategory(@Request() req, @Body() createCategoryRequestDto: CreateCategoryRequestDto): Promise<Category> {
-    const restaurantId = req.user.restaurantId;
-
-    return await this.categoriesService.createCategory(restaurantId, createCategoryRequestDto);
-  }
-
-  @Put(':category_id')
-  async updateCategory(
+  async createCategory(
     @Request() req,
-    @Param('category_id') categoryId: string,
-    @Body() updateCategoryRequestDto: UpdateCategoryRequestDto,
-  ): Promise<Category> {
-    const restaurantId = req.user.restaurantId;
-
-    return await this.categoriesService.updateCategory(restaurantId, categoryId, updateCategoryRequestDto);
+    @Body() createCategoryRequestDto: CreateCategoryRequestDto,
+  ): Promise<{ categoryId: string }> {
+    if (req.user.type === 'login') {
+      const restaurantId = req.user.restaurantId;
+      const categoryId = await this.categoriesService.createCategory(restaurantId, createCategoryRequestDto);
+      return { categoryId };
+    }
+    throw new UnauthorizedException('로그인이 필요한 기능입니다.');
   }
 
-  @Delete(':category_id')
-  async deleteCategory(@Request() req, @Param('category_id') categoryId: string): Promise<void> {
-    const restaurantId = req.user.restaurantId;
-
-    return await this.categoriesService.deleteCategory(restaurantId, categoryId);
+  @Delete(':categoryId')
+  async deleteCategory(@Request() req, @Param('categoryId') categoryId: string): Promise<void> {
+    if (req.user.type === 'login') {
+      const restaurantId = req.user.restaurantId;
+      return await this.categoriesService.deleteCategory(restaurantId, categoryId);
+    }
+    throw new UnauthorizedException('로그인이 필요한 기능입니다.');
   }
 }
